requirements.txt
torch==2.0.0+cu118
torchvision==0.15.1+cu118
onnx==1.14.0
onnxruntime==1.15.1
transformers==4.46.3
mmcv==2.2.0
numpy==1.24.2
pandas==2.0.3
matplotlib==3.7.1
Pillow==9.4.0
scipy==1.10.1
pycocotools==2.0.7
tqdm 
PyYAML==6.0
shapely==2.0.7